+++ backend/src/main/java/com/zzang/chongdae/offering/util/httpclient/crawler/JsoupHtmlCrawler.java
@@ -1,4 +1,4 @@
-package com.zzang.chongdae.offering.util.httpclient;
+package com.zzang.chongdae.offering.util.httpclient.crawler;
 
 import java.io.IOException;
 import java.net.MalformedURLException;
